Building NetworkCompose on top of NetworkBuilder to support multiple call with priority
