+++ .build/workspace-state.json
@@ -1,0 +1,11 @@
+{
+  "object" : {
+    "artifacts" : [
+
+    ],
+    "dependencies" : [
+
+    ]
+  },
+  "version" : 6
+}
