+++ Example/Example/DemoScenario.swift
@@ -1,0 +1,18 @@
+//
+//  DemoScenario.swift
+//  Example
+//
+//  Created by Hoang Nguyen on 28/11/23.
+//
+
+import Foundation
+
+enum DemoScenario: String {
+    case defaultRequest = "Demo default request"
+    case reAuthentication = "Demo reauthentication request"
+    case enabledSSLPinning = "Demo enabled SSL Pinning request"
+    case networkMetricReport = "Demo collect metric report"
+    case smartRetry = "Demo smart retry request"
+    case multipleRequetWithPriority = "Demo multiple requests with priority"
+    case supportAutomationTest = "Demo suppport automation testing"
+}
