+++ Sources/NetworkCompose/src/Network/NetworkBuilder.swift
@@ -7,15 +7,23 @@
 
 import Foundation
 
-public class NetworkBuilder<SessionType: NetworkSession>: NetworkCommonSettings<SessionType> {
+/// A builder class responsible for configuring and creating network-related functionality.
+public class NetworkBuilder<SessionType: NetworkSession>: NetworkBuilderSettings<SessionType> {
+    /// The service responsible for re-authentication.
     private var reAuthService: ReAuthenticationService?
+
+    /// The queue used to run re-authentication operations.
     private var operationQueue: OperationQueueManagerInterface = DefaultOperationQueueManager.serialOperationQueue
 
+    /// Initializes a new instance of `NetworkBuilder`.
+    ///
+    /// - Parameters:
+    ///   - baseURL: The base URL for network requests.
+    ///   - session: The network session to be used. Defaults to `URLSession.shared`.
     public required init(baseURL: URL,
                          session: SessionType = URLSession.shared)
     {
-        super.init(baseURL: baseURL,
-                   session: session)
+        super.init(baseURL: baseURL, session: session)
     }
 
     /// Sets the re-authentication service for the builder.
@@ -30,7 +38,7 @@ public class NetworkBuilder<SessionType: NetworkSession>: NetworkCommonSettings<
 
     /// Sets the operation-queue service for the builder.
     ///
-    /// - Parameter operationQueue: The queue run re-authentication operation
+    /// - Parameter operationQueue: The queue to run re-authentication operations.
     /// - Returns: The builder instance for method chaining.
     @discardableResult
     public func reAuthenOperationQueue(_ operationQueue: OperationQueueManagerInterface) -> Self {
@@ -67,24 +75,30 @@ public class NetworkBuilder<SessionType: NetworkSession>: NetworkCommonSettings<
         return self
     }
 
-    /// Builds and returns an instance conforming to `NetworkCoordinatorInterface` based on the configured strategies.
+    /// Builds and returns an instance conforming to `NetworkRouterInterface` based on the configured strategies.
     ///
-    /// This method creates either a `NetworkCoordinator` or a `NetworkMocker` based on the specified strategies.
+    /// This method creates either a `NetworkNavigator` or a `NetworkMocker` based on the specified strategies.
     ///
-    /// - Returns: An instance conforming to `NetworkCoordinatorInterface`.
-    private func build() -> NetworkCoordinatorInterface {
-        guard case let .enabled(mockDataType) = automationMode else {
+    /// - Returns: An instance conforming to `NetworkRouterInterface`.
+    public func build() -> NetworkRouterInterface {
+        if case let .enabled(mockDataType) = automationMode {
+            return NetworkMocker(
+                baseURL: baseURL,
+                session: session,
+                reAuthService: reAuthService,
+                executionQueue: executionQueue,
+                observationQueue: observationQueue,
+                loggerInterface: createLogger(),
+                mockDataType: mockDataType
+            )
+        } else {
             var storageService: StorageService?
-
             if case .enabled = recordResponseMode {
                 storageService = StorageServiceProvider(loggerInterface: createLogger(),
                                                         executionQueue: executionQueue)
             }
-            if let session = try? createNetworkSession() {
-                self.session = session
-            }
-
-            return NetworkCoordinator(
+            if let session = try? createNetworkSession() { self.session = session }
+            return NetworkRouter(
                 baseURL: baseURL,
                 session: session,
                 reAuthService: reAuthService,
@@ -96,45 +110,5 @@ public class NetworkBuilder<SessionType: NetworkSession>: NetworkCommonSettings<
                 loggerInterface: createLogger()
             )
         }
-
-        return NetworkMocker(
-            baseURL: baseURL,
-            session: session,
-            reAuthService: reAuthService,
-            executionQueue: executionQueue,
-            observationQueue: observationQueue,
-            loggerInterface: createLogger(),
-            mockDataType: mockDataType
-        )
-    }
-}
-
-public extension NetworkBuilder {
-    func request<RequestType: RequestInterface>(
-        _ request: RequestType,
-        andHeaders headers: [String: String] = [:],
-        retryPolicy: RetryPolicy = .none,
-        completion: @escaping (Result<RequestType.SuccessType, NetworkError>) -> Void
-    ) {
-        build().request(request, andHeaders: headers, retryPolicy: retryPolicy, completion: completion)
-    }
-
-    func upload<RequestType: RequestInterface>(
-        _ request: RequestType,
-        andHeaders headers: [String: String] = [:],
-        fromFile fileURL: URL,
-        retryPolicy: RetryPolicy = .none,
-        completion: @escaping (Result<RequestType.SuccessType, NetworkError>) -> Void
-    ) {
-        build().upload(request, andHeaders: headers, fromFile: fileURL, retryPolicy: retryPolicy, completion: completion)
-    }
-
-    func download<RequestType: RequestInterface>(
-        _ request: RequestType,
-        andHeaders headers: [String: String] = [:],
-        retryPolicy: RetryPolicy = .none,
-        completion: @escaping (Result<RequestType.SuccessType, NetworkError>) -> Void
-    ) {
-        build().download(request, andHeaders: headers, retryPolicy: retryPolicy, completion: completion)
     }
 }
