--- Sources/NetworkCompose/src/Network/NetworkControllerInterface.swift
@@ -1,96 +1,0 @@
-//
-//  NetworkControllerInterface.swift
-//  NetworkCompose
-//
-//  Created by Hoang Nguyen on 11/11/23.
-//
-
-import Foundation
-
-public protocol NetworkControllerInterface: AnyObject {
-    /// Sends a network request and executes the completion handler with the result.
-    ///
-    /// - Parameters:
-    ///   - request: The network request to be performed.
-    ///   - headers: Additional headers to be included in the request.
-    ///   - retryPolicy: The retry policy for the network request.
-    ///   - completion: The completion handler to be called with the result.
-    ///
-    /// - Note: Use this method for non-async network requests or when compatibility with earlier iOS versions is required.
-    func request<RequestType: RequestInterface>(
-        _ request: RequestType,
-        andHeaders headers: [String: String],
-        retryPolicy: RetryPolicy,
-        completion: @escaping (Result<RequestType.SuccessType, NetworkError>) -> Void
-    )
-
-    /// Uploads a file using a network request and executes the completion handler with the result.
-    ///
-    /// - Parameters:
-    ///   - request: The network request to be performed.
-    ///   - headers: Additional headers to be included in the request.
-    ///   - fileURL: The URL of the file to be uploaded.
-    ///   - retryPolicy: The retry policy for the network request.
-    ///   - completion: The completion handler to be called with the result.
-    func upload<RequestType: RequestInterface>(
-        _ request: RequestType,
-        andHeaders headers: [String: String],
-        fromFile fileURL: URL,
-        retryPolicy: RetryPolicy,
-        completion: @escaping (Result<RequestType.SuccessType, NetworkError>) -> Void
-    )
-
-    /// Downloads a file using a network request and executes the completion handler with the result.
-    ///
-    /// - Parameters:
-    ///   - request: The network request to be performed.
-    ///   - headers: Additional headers to be included in the request.
-    ///   - retryPolicy: The retry policy for the network request.
-    ///   - completion: The completion handler to be called with the result.
-    func download<RequestType: RequestInterface>(
-        _ request: RequestType,
-        andHeaders headers: [String: String],
-        retryPolicy: RetryPolicy,
-        completion: @escaping (Result<RequestType.SuccessType, NetworkError>) -> Void
-    )
-}
-
-public extension NetworkControllerInterface {
-    func request<RequestType: RequestInterface>(
-        _ request: RequestType,
-        andHeaders headers: [String: String] = [:],
-        retryPolicy: RetryPolicy = .none,
-        completion: @escaping (Result<RequestType.SuccessType, NetworkError>) -> Void
-    ) {
-        self.request(request,
-                     andHeaders: headers,
-                     retryPolicy: retryPolicy,
-                     completion: completion)
-    }
-
-    func upload<RequestType: RequestInterface>(
-        _ request: RequestType,
-        andHeaders headers: [String: String] = [:],
-        fromFile fileURL: URL,
-        retryPolicy: RetryPolicy = .none,
-        completion: @escaping (Result<RequestType.SuccessType, NetworkError>) -> Void
-    ) {
-        upload(request,
-               andHeaders: headers,
-               fromFile: fileURL,
-               retryPolicy: retryPolicy,
-               completion: completion)
-    }
-
-    func download<RequestType: RequestInterface>(
-        _ request: RequestType,
-        andHeaders headers: [String: String] = [:],
-        retryPolicy: RetryPolicy = .none,
-        completion: @escaping (Result<RequestType.SuccessType, NetworkError>) -> Void
-    ) {
-        download(request,
-                 andHeaders: headers,
-                 retryPolicy: retryPolicy,
-                 completion: completion)
-    }
-}
