+++ Sources/NetworkCompose/src/Network/OperationQueue/CustomOperation.swift
@@ -7,6 +7,31 @@
 
 import Foundation
 
+/// `ClosureCustomOperation` is a final class that extends `CustomOperation` to provide a customizable operation
+/// using a closure. This class allows you to encapsulate a specific task or set of tasks within a closure
+/// and execute it asynchronously as part of an operation queue.
+final class ClosureCustomOperation: CustomOperation {
+    typealias Closure = (ClosureCustomOperation) -> Void
+    private let closure: Closure
+
+    /// Initializes the operation with the given closure.
+    ///
+    /// - Parameter closure: The closure to be executed when the operation is started.
+    init(closure: @escaping Closure) {
+        self.closure = closure
+    }
+
+    /// The main execution point for the operation.
+    ///
+    /// This method is called when the operation is ready to execute.
+    override func main() {
+        guard !isCancelled else {
+            return
+        }
+        closure(self)
+    }
+}
+
 class CustomOperation: Operation {
     var validOperation: Bool = true
