+++ Sources/NetworkCompose/src/NetworkMocker/EndpointExpectation.swift
@@ -7,37 +7,50 @@
 
 import Foundation
 
+/// A struct representing an expectation for a network endpoint, including a unique key and the expected response.
 public struct EndpointExpectation {
-    public let name: String
-    public let path: String
-    public let method: NetworkMethod
+    /// The unique key associated with the network request.
+    public let uniqueKey: UniqueKey
+
+    /// The expected response for the network request.
     public let response: Response
 
+    /// The possible responses for the network request.
     public enum Response {
+        /// Represents a failure response with a specific network error.
         case failure(NetworkError)
+
+        /// Represents a success response with a decodable result.
         case successResponse(Decodable)
     }
 
-    public init(name: String,
-                path: String,
-                method: NetworkMethod,
-                response: Response)
-    {
-        self.name = name
-        self.path = path
-        self.method = method
+    /// Initializes an `EndpointExpectation` instance with the specified components.
+    ///
+    /// - Parameters:
+    ///   - path: The path of the network request.
+    ///   - method: The HTTP method of the network request.
+    ///   - queryParameters: Optional query parameters for the network request. Default is `nil`.
+    ///   - response: The expected response for the network request.
+    public init(path: String, method: NetworkMethod, queryParameters: [String: Any]? = nil, response: Response) {
+        uniqueKey = UniqueKey(path: path, method: method.rawValue, queryParameters: queryParameters)
         self.response = response
     }
 
-    public func isSameRequest<RequestType>(
-        _ request: RequestType
-    ) -> Bool where RequestType: RequestInterface {
-        return path == request.path && method == request.method
+    /// Compares the current expectation with a given network request to check if they represent the same request.
+    ///
+    /// - Parameter request: The network request to compare.
+    /// - Returns: `true` if the requests are the same; otherwise, `false`.
+    public func isSameRequest<RequestType>(_ request: RequestType) -> Bool where RequestType: RequestInterface {
+        let compare = UniqueKey(path: request.path, method: request.method.rawValue, queryParameters: request.queryParameters)
+        return uniqueKey.key == compare.key
     }
 
-    public func getResponse<RequestType>(
-        _: RequestType
-    ) throws -> RequestType.SuccessType where RequestType: RequestInterface {
+    /// Retrieves the expected success response for a given network request.
+    ///
+    /// - Parameter request: The network request for which the response is expected.
+    /// - Returns: The expected success response.
+    /// - Throws: An error if the response type is not the same as the expected success type.
+    public func getResponse<RequestType>(_: RequestType) throws -> RequestType.SuccessType where RequestType: RequestInterface {
         switch response {
         case let .failure(error):
             throw error
@@ -51,7 +64,13 @@ public struct EndpointExpectation {
 }
 
 extension EndpointExpectation: Equatable {
+    /// Checks if two `EndpointExpectation` instances are equal by comparing their unique keys.
+    ///
+    /// - Parameters:
+    ///   - lhs: The left-hand side `EndpointExpectation`.
+    ///   - rhs: The right-hand side `EndpointExpectation`.
+    /// - Returns: `true` if the unique keys are equal; otherwise, `false`.
     public static func == (lhs: EndpointExpectation, rhs: EndpointExpectation) -> Bool {
-        return lhs.name == rhs.name && lhs.path == rhs.path && lhs.method == rhs.method
+        return lhs.uniqueKey == rhs.uniqueKey
     }
 }
